# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KingSaint Capital - Premium Crypto Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    User:
      type: object
      required: [id, email, fullName, role, status, balance, totalInvested, totalProfit, referralCode, createdAt]
      properties:
        id:
          type: number
        email:
          type: string
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, pending]
        balance:
          type: number
        totalInvested:
          type: number
        totalProfit:
          type: number
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        country:
          type: string
        avatarUrl:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
...